Remove T-SQL tutorial
The documentation of "handling T-SQL" was more like an
introduction or a tutorial than technical documentation.

This will be uninteresting to the readers who already know T-SQL,
and it will be too incomplete to be useful to beginners.

There is already an example of a T-SQL procedure in the FAQ,
which should be enough to showcase what is possible.

To be on the safe side, extend the "migration" documentation
to state that most T-SQL, including functions and procedures,
is supported.

The reference to the removed page was removed from the FAQ.

In passing, add an HTML "abbr" tag to explain the
abbreviation DDL those who don't know it.

This closes #119.

Signed-off-by: Laurenz Albe <laurenz.albe@cybertec.at>