Replace "babelfish_money" documentation with a link
Since this is just a copy of the 2ndQuadrant software, it
is silly to add usage instructions and internal details.

Rather, point to the original documentation.

This closes #103.

Not that this commit doesn't touch any of the other problems
in the data type documentation, in order to reduce merge conflicts
with a prospective patch that addresses #102 once we have the
required information.

Signed-off-by: Laurenz Albe <laurenz.albe@cybertec.at>