Add a link to finish Babelfish setup
When people want to install Babelfish, they will read the
"compilation from source" page.  However, that page does not
mention that you need to initialze Babelfish.
The reader is left thinking that Babelfish is ready to use
once the server is started.

Add a link to the "single-db vs. multi-db setup" page.

Signed-off-by: Laurenz Albe <laurenz.albe@cybertec.at>