Unsupported Feature Section - Website Reformatting (#234)
This PR adds the _limitations folder and _config update file.  The changes will add an 'Unsupported Features' drop-down menu, and move the content related to unsupported features into that section.  In the first PR, I'm re-using the Unsupported features file that we already have; in a future PR, the unsupported content will be broken into separate files for each version we release.